requirements.txt
fastapi
uvicorn
pydantic

langchain-core
langchain-openai
langchain-community
langchain-text-splitters
langchain-ollama
langchain_chroma

chromadb

python-dotenv
openai

pytest
httpx